(#240) Refactor of pr/issue comments
*  Factor out get_issue_comments to base class
*  Make filtering for both pr/issue same
