+++ ogr/services/base.py
@@ -20,11 +20,14 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-from typing import List, Optional, Match, Any
+from typing import List, Optional, Match, Any, TypeVar
 
-from ogr.abstract import GitService, GitProject, PRComment, GitUser
+from ogr.abstract import GitService, GitProject, PRComment, GitUser, IssueComment
 from ogr.parsing import parse_git_repo
 from ogr.utils import search_in_comments, filter_comments
+
+
+Comment = TypeVar("Comment", IssueComment, PRComment)
 
 
 class BaseGitService(GitService):
@@ -45,8 +48,18 @@ class BaseGitProject(GitProject):
         """
         return f"{self.namespace}/{self.repo}"
 
+    @staticmethod
+    def __get_comments(
+        comments: List[Comment], filter_regex: str, reverse: bool, author: str
+    ) -> List[Comment]:
+        if reverse:
+            comments.reverse()
+        if filter_regex or author:
+            comments = filter_comments(comments, filter_regex, author)
+        return comments
+
     def get_pr_comments(
-        self, pr_id, filter_regex: str = None, reverse: bool = False
+        self, pr_id, filter_regex: str = None, reverse: bool = False, author: str = None
     ) -> List[PRComment]:
         """
         Get list of pull-request comments.
@@ -54,14 +67,30 @@ class BaseGitProject(GitProject):
         :param pr_id: int
         :param filter_regex: filter the comments' content with re.search
         :param reverse: reverse order of comments
+        :param author: filter comments by author
         :return: [PRComment]
         """
         all_comments = self._get_all_pr_comments(pr_id=pr_id)
-        if reverse:
-            all_comments.reverse()
-        if filter_regex:
-            all_comments = filter_comments(all_comments, filter_regex)
-        return all_comments
+        return self.__get_comments(all_comments, filter_regex, reverse, author)
+
+    def get_issue_comments(
+        self,
+        issue_id,
+        filter_regex: str = None,
+        reverse: bool = False,
+        author: str = None,
+    ) -> List[IssueComment]:
+        """
+        Get list of issue comments.
+
+        :param pr_id: int
+        :param filter_regex: filter the comments' content with re.search
+        :param reverse: reverse order of comments
+        :param author: filter comments by author
+        :return: [PRComment]
+        """
+        all_comments = self._get_all_issue_comments(issue_id=issue_id)
+        return self.__get_comments(all_comments, filter_regex, reverse, author)
 
     def search_in_pr(
         self,
