+++ ogr/services/gitlab/project.py
@@ -21,7 +21,6 @@
 # SOFTWARE.
 
 import logging
-import re
 from typing import List, Optional, Dict, Set
 
 import gitlab
@@ -210,21 +209,15 @@ class GitlabProject(BaseGitProject):
             if member.access_level in access_levels
         ]
 
-    def get_issue_comments(
-        self, issue_id, filter_regex: str = None, reverse: bool = False
-    ) -> List["IssueComment"]:
+    def _get_all_issue_comments(self, issue_id) -> List["IssueComment"]:
         issue = self.gitlab_repo.issues.get(issue_id)
-        if not filter_regex:
-            filter_regex = r".*"
 
         comments = [
             self._issuecomment_from_gitlab_object(raw_comment)
             for raw_comment in issue.notes.list()
-            if re.search(filter_regex, raw_comment.body)
         ]
 
-        if not reverse:
-            comments.reverse()
+        comments.reverse()
 
         return comments
 
@@ -436,13 +429,6 @@ class GitlabProject(BaseGitProject):
             {"title": title, "description": description}
         )
         return self._issue_from_gitlab_object(issue)
-
-    def _get_all_issue_comments(self, issue_id: int) -> List[IssueComment]:
-        issue = self.gitlab_repo.issues.get(issue_id)
-        return [
-            self._issuecomment_from_gitlab_object(raw_comment)
-            for raw_comment in issue.notes.list()
-        ]
 
     def issue_comment(self, issue_id: int, body: str) -> IssueComment:
         """
